+++ test/src/redirectors_test.dart
@@ -1,16 +1,5 @@
-// Copyright (c) 2022, Minh Vuong
-// https://github.com/vuongxuongminh
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: lines_longer_than_80_chars
-
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:go_router/go_router.dart';
 import 'package:go_router_plus/go_router_plus.dart';
 import 'package:mockito/annotations.dart';