+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+3.0.0
+-----
+
++ **BC:**  replace method [Screen.parentNavigatorKey()] with [Screen.parentNavigatorKey] getter.
++ **BC:** replace method [ShellScreen.navigatorKey()] with [ShellScreen.navigatorKey] getter.
++ [ShellScreen] add getter observers.
++ Bump version of Go Router version `^6.2.0`.
++ Bump version of `very_good_analysis` to `^4.0.0`.
++ Refactor codebase.
++ Export package `package:go_router/go_router.dart` for dev experience, so we don't need to require both go_router and go_router_plus.
+
 2.1.0
 -----
 + Add support Go Router 6.